requirements.txt
Django==3.0.8
djangorestframework==3.11.0
jwt==1.0.0
PyJWT==1.7.1